Creation of all modals (old, not components yet)
